+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/actions/FlexibleSearchToggleInEditorResultsAction.kt
@@ -20,14 +20,15 @@ package com.intellij.idea.plugin.hybris.flexibleSearch.actions
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.flexibleSearch.editor.flexibleSearchSplitEditor
+import com.intellij.idea.plugin.hybris.project.utils.Plugin
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.ToggleAction
 
 class FlexibleSearchToggleInEditorResultsAction : ToggleAction(
-    message("hybris.fxs.actions.in_editor_results"),
-    message("hybris.fxs.actions.in_editor_results.description"),
-    HybrisIcons.FlexibleSearch.TOGGLE_IN_EDITOR_RESULTS
+    message("hybris.actions.in_editor_results"),
+    message("hybris.actions.in_editor_results.description"),
+    HybrisIcons.Actions.TOGGLE_IN_EDITOR_RESULTS
 ) {
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT
@@ -39,4 +40,8 @@ class FlexibleSearchToggleInEditorResultsAction : ToggleAction(
         e.flexibleSearchSplitEditor()?.inEditorResults = state
     }
 
+    override fun update(e: AnActionEvent) {
+        super.update(e)
+        e.presentation.isVisible = e.presentation.isVisible && Plugin.GRID.isActive()
+    }
 }