+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 - Improved _numbers_ editors in the `In-Editor Parameters` view [#1453](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1453)
 - Automatically Escape `'` in the _text_ editors in the `In-Editor Parameters` view [#1454](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1454)
 
+### `Polyglot Query` enhancements
+
+- Introduced `In-Editor Results` view [#1462](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1462)
+
 ### `Groovy` enhancements
 - Introduced `batch` execution mode for Groovy files [#1450](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1450)